+++ src/modules/animation/animal/bird/anim/Run.cpp
@@ -10,7 +10,7 @@ namespace animation {
 namespace animal {
 namespace bird {
 namespace run {
-void update(float animTime, float velocity, BirdSkeleton &skeleton, const BirdSkeletonAttribute &skeletonAttr) {
+void update(double animTime, double velocity, BirdSkeleton &skeleton, const BirdSkeletonAttribute &skeletonAttr) {
 	const float timeFactor = skeletonAttr.runTimeFactor;
 	const float sine = glm::sin(animTime * timeFactor);
 	const float cosine = glm::cos(animTime * timeFactor);
@@ -18,7 +18,7 @@ void update(float animTime, float velocity, BirdSkeleton &skeleton, const BirdSk
 	const float movement = sine * 0.35f;
 	const glm::vec2 headLook(glm::cos(animTime) * 0.05f + glm::radians(10.0f), sine * 0.1f);
 
-	velocity = glm::clamp(velocity / 10.0f, 0.1f, 2.5f);
+	velocity = glm::clamp((float)(velocity / 10.0), 0.1f, 2.5f);
 
 	Bone &head = skeleton.headBone(skeletonAttr);
 	head.translation = glm::vec3(0.0f, skeletonAttr.headY + cosine * 1.3f, 0.0f);
