+++ src/tools/mapview/MapView.h
@@ -53,8 +53,8 @@ protected:
 	bool _freelook = false;
 	bool _updateWorld = true;
 	int _drawCallsWorld = 0;
-	float _timeScaleFactor = 480.0f;
-	float _worldTime = 0.0f;
+	double _timeScaleFactor = 480.0;
+	double _worldTime = 0.0;
 
 	core::VarPtr _meshSize;
